backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.0
pymongo==4.6.3
dnspython==2.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
itsdangerous==2.2.0
tensorflow==2.16.2
opencv-python-headless==4.8.1.78
Pillow==10.1.0
